requirements.txt
google-cloud-vision==3.4.4
requests==2.31.0
vertexai==1.43.0
openai==1.12.0
google-auth==2.27.0
python-dotenv==1.0.1
